pyproject.toml
[project]
name = "rag-framework"
version = "0.1.0"
description = "RAG Pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate==1.13.0",
    "aio-pika==9.6.2",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.5",
    "aiormq==6.9.4",
    "aiosignal==1.4.0",
    "alembic==1.18.3",
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "appdirs==1.4.4",
    "async-timeout==5.0.1",
    "asyncpg==0.31.0",
    "attrs==26.1.0",
    "backoff==2.2.1",
    "certifi==2026.1.4",
    "charset-normalizer==3.4.4",
    "click==8.3.3",
    "colorama==0.4.6",
    "dataclasses-json==0.6.7",
    "datasets==4.8.5",
    "dill==0.4.1",
    "diskcache==5.6.3",
    "distro==1.9.0",
    "docstring-parser==0.18.0",
    "et-xmlfile==2.0.0",
    "exceptiongroup==1.3.1",
    "fast-depends==3.0.8",
    "fastapi==0.136.1",
    "faststream==0.6.7",
    "filelock==3.29.0",
    "frozenlist==1.8.0",
    "fsspec==2026.2.0",
    "googleapis-common-protos==1.72.0",
    "greenlet==3.3.1",
    "grpcio==1.78.1",
    "h11==0.16.0",
    "hf-xet==1.4.3",
    "httpcore==1.0.9",
    "httptools==0.7.1",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==1.12.0",
    "idna==3.11",
    "importlib-metadata==8.7.1",
    "instructor==1.15.1",
    "jinja2==3.1.6",
    "jiter==0.13.0",
    "joblib==1.5.3",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.2.10",
    "langchain-community==0.3.31",
    "langchain-core==1.2.13",
    "langchain-huggingface==1.2.1",
    "langchain-openai==1.1.10",
    "langchain-text-splitters==1.1.1",
    "langfuse==4.2.0",
    "langgraph==1.0.8",
    "langgraph-checkpoint==4.0.0",
    "langgraph-prebuilt==1.0.7",
    "langgraph-sdk==0.3.7",
    "langsmith==0.7.4",
    "llmlingua==0.2.2",
    "mako==1.3.10",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.3",
    "marshmallow==3.26.2",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.7.1",
    "multiprocess==0.70.19",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.4.2",
    "nltk==3.9.4",
    "numpy==2.2.6",
    "openai==2.33.0",
    "openpyxl==3.1.5",
    "opentelemetry-api==1.41.0",
    "opentelemetry-exporter-otlp-proto-common==1.41.0",
    "opentelemetry-exporter-otlp-proto-grpc==1.41.0",
    "opentelemetry-exporter-otlp-proto-http==1.41.0",
    "opentelemetry-proto==1.41.0",
    "opentelemetry-sdk==1.41.0",
    "opentelemetry-semantic-conventions==0.62b0",
    "orjson==3.11.7",
    "ormsgpack==1.12.2",
    "packaging==26.0",
    "pamqp==3.3.0",
    "pandas==2.3.3",
    "pgvector==0.4.2",
    "pillow==12.2.0",
    "propcache==0.4.1",
    "protobuf==6.33.5",
    "psutil==7.2.2",
    "psycopg==3.3.2",
    "psycopg-binary==3.3.2",
    "psycopg2-binary==2.9.12",
    "pyarrow==24.0.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.14.0",
    "pygments==2.20.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pytz==2026.1.post1",
    "pyyaml==6.0.3",
    "ragas==0.4.3",
    "redis==7.4.0",
    "regex==2026.4.4",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rich==14.3.4",
    "safetensors==0.7.0",
    "scikit-learn==1.7.2",
    "scikit-network==0.33.5",
    "scipy==1.15.3",
    "sentence-transformers==5.4.1",
    "setuptools==81.0.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.46",
    "starlette==1.0.0",
    "sympy==1.14.0",
    "tenacity==9.1.4",
    "threadpoolctl==3.6.0",
    "tiktoken==0.12.0",
    "tokenizers==0.22.2",
    "tomli==2.4.0",
    "torch==2.11.0",
    "tqdm==4.67.3",
    "transformers==5.7.0",
    "typer==0.25.0",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "tzdata==2026.2",
    "urllib3==2.6.3",
    "uuid-utils==0.14.0",
    "uvicorn==0.46.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "websockets==16.0",
    "wrapt==1.17.3",
    "xxhash==3.6.0",
    "yarl==1.23.0",
    "zipp==3.23.0",
    "zstandard==0.25.0",
]
